docs: bug-batch classification ledger (issues #1685 #1659 #1666 #1664 #1690)
